fix ctr+c handling and small refactoring
