+++ main.go
@@ -5,12 +5,10 @@ import (
 	"flag"
 	"fmt"
 	"log/slog"
-	"net"
 	"net/http"
 	"os"
 	"os/exec"
 	"os/signal"
-	"strconv"
 	"strings"
 	"syscall"
 
@@ -49,54 +47,24 @@ func main() {
 	if *isDebug {
 		lvl = slog.LevelDebug
 	}
+
+	whoami := &proxyInstance{
+		Host:    "localhost",
+		Port:    *httPort,
+		Title:   *pageTitle,
+		Command: *targetCommand,
+	}
+	selector := &instanceSelector{
+		current: whoami,
+	}
+
 	reclog := slog.New(nanny.NewLogHandler(rec, logHandler, lvl))
 	slog.SetDefault(reclog)
 	options := nanny.BrowserOptions{
-		PageSize:  100,
-		PageTitle: *pageTitle,
-		BeforeHTMLTableFunc: func() string {
-			instances, err := readRegistryEntries()
-			sb := strings.Builder{}
-			if err != nil {
-				sb.WriteString("<mark>")
-				sb.WriteString(err.Error())
-				sb.WriteString("</mark>")
-				return sb.String()
-			}
-			sb.WriteString("<select id=\"instance-selector\">")
-
-			// Get current host:port
-			currentHostPort := "localhost:" + strconv.Itoa(*httPort)
-
-			for _, i := range instances {
-				instanceURL := "http://" + i.Host + ":" + strconv.Itoa(i.Port)
-				selected := ""
-				if i.Host+":"+strconv.Itoa(i.Port) == currentHostPort {
-					selected = " selected"
-				}
-				sb.WriteString("<option value=\"" + instanceURL + "\"" + selected + ">")
-				sb.WriteString(i.Title + " (" + i.Host + ":" + strconv.Itoa(i.Port) + ")")
-				sb.WriteString("</option>")
-			}
-
-			sb.WriteString("</select>")
-
-			return sb.String()
-		},
-		EndHTMLHeadFunc: func() string {
-			return `
-			<script>
-			document.addEventListener('DOMContentLoaded', function() {
-				var selector = document.getElementById('instance-selector');
-				if (selector) {
-					selector.addEventListener('change', function() {
-						window.location.href = this.value;
-					});
-				}
-			});
-			</script>
-			`
-		},
+		PageSize:            100,
+		PageTitle:           *pageTitle,
+		BeforeHTMLTableFunc: selector.beforeTableHTML,
+		EndHTMLHeadFunc:     selector.endHeadHTML,
 	}
 	http.Handle("/", nanny.NewBrowser(rec, options))
 
@@ -127,41 +95,11 @@ func main() {
 		<-sigChan
 		slog.Info("received termination signal, shutting down")
 		cancel()
+		abort(whoami, 0)
 	}()
-
-	whoami := proxyInstance{
-		Host:    "localhost",
-		Port:    *httPort,
-		Title:   *pageTitle,
-		Command: *targetCommand,
-	}
 
 	// serve nanny
-	go func() {
-		if err := addToOrRemoveFromRegistry(whoami, false); err != nil {
-			slog.Error("failed to add to registry", "error", err)
-		} else {
-			slog.Debug("added to registry", "instance", whoami)
-		}
-		// use the given port to listen, fall back to a free port if it is already in use
-		if err := http.ListenAndServe(net.JoinHostPort("localhost", strconv.Itoa(whoami.Port)), nil); err != nil {
-			if strings.Contains(err.Error(), "bind: address already in use") {
-				// try with a different port
-				newPort, err := GetFreePort()
-				if err == nil {
-					whoami.Port = newPort
-					// add new port to registry
-					addToOrRemoveFromRegistry(whoami, false)
-					http.ListenAndServe(net.JoinHostPort("localhost", strconv.Itoa(whoami.Port)), nil)
-				} else {
-					slog.Error("failed to get free port, cannot start log service", "error", err)
-				}
-			} else {
-				slog.Error("failed to start HTTP log service", "error", err)
-			}
-		}
-		addToOrRemoveFromRegistry(whoami, true)
-	}()
+	go registerAndStartListener(whoami)
 
 	// client -> proxy -> target
 	go func() {
@@ -189,24 +127,9 @@ func main() {
 	}
 }
 
-func abort(pi proxyInstance, code int) {
+func abort(pi *proxyInstance, code int) {
 	if err := addToOrRemoveFromRegistry(pi, true); err != nil {
 		slog.Error("failed to remote from registry on abort", "error", err)
 	}
 	os.Exit(code)
-}
-
-// GetFreePort asks the kernel for a free open port that is ready to use.
-func GetFreePort() (int, error) {
-	addr, err := net.ResolveTCPAddr("tcp", "localhost:0")
-	if err != nil {
-		return 0, err
-	}
-
-	l, err := net.ListenTCP("tcp", addr)
-	if err != nil {
-		return 0, err
-	}
-	defer l.Close()
-	return l.Addr().(*net.TCPAddr).Port, nil
 }
